+++ crates/biome_js_analyze/src/options.rs
@@ -2,6 +2,7 @@
 
 use crate::analyzers::nursery::use_filenaming_convention::FilenamingConventionOptions;
 use crate::semantic_analyzers::correctness::use_exhaustive_dependencies::HooksOptions;
+use crate::semantic_analyzers::correctness::use_hook_at_top_level::DeprecatedHooksOptions;
 use crate::semantic_analyzers::style::no_restricted_globals::RestrictedGlobalsOptions;
 use crate::semantic_analyzers::style::use_naming_convention::NamingConventionOptions;
 use crate::{
@@ -24,8 +25,10 @@ pub enum PossibleOptions {
     Complexity(ComplexityOptions),
     /// Options for `useFilenamingConvention` rule
     FilenamingConvention(FilenamingConventionOptions),
-    /// Options for `useExhaustiveDependencies` and `useHookAtTopLevel` rule
+    /// Options for `useExhaustiveDependencies` rule
     Hooks(HooksOptions),
+    /// Deprecated options for `useHookAtTopLevel` rule
+    DeprecatedHooks(DeprecatedHooksOptions),
     /// Options for `useNamingConvention` rule
     NamingConvention(NamingConventionOptions),
     /// Options for `noRestrictedGlobals` rule
@@ -50,7 +53,7 @@ impl PossibleOptions {
                 };
                 RuleOptions::new(options)
             }
-            "useExhaustiveDependencies" | "useHookAtTopLevel" => {
+            "useExhaustiveDependencies" => {
                 let options = match self {
                     PossibleOptions::Hooks(options) => options.clone(),
                     _ => HooksOptions::default(),
@@ -61,6 +64,13 @@ impl PossibleOptions {
                 let options = match self {
                     PossibleOptions::FilenamingConvention(options) => options.clone(),
                     _ => FilenamingConventionOptions::default(),
+                };
+                RuleOptions::new(options)
+            }
+            "useHookAtTopLevel" => {
+                let options = match self {
+                    PossibleOptions::DeprecatedHooks(options) => options.clone(),
+                    _ => DeprecatedHooksOptions::default(),
                 };
                 RuleOptions::new(options)
             }
@@ -103,9 +113,11 @@ impl Deserializable for PossibleOptions {
             }
             "noRestrictedGlobals" => Deserializable::deserialize(value, "options", diagnostics)
                 .map(Self::RestrictedGlobals),
-            "useExhaustiveDependencies" | "useHookAtTopLevel" => {
+            "useExhaustiveDependencies" => {
                 Deserializable::deserialize(value, "options", diagnostics).map(Self::Hooks)
             }
+            "useHookAtTopLevel" => Deserializable::deserialize(value, "options", diagnostics)
+                .map(Self::DeprecatedHooks),
             "useFilenamingConvention" => Deserializable::deserialize(value, "options", diagnostics)
                 .map(Self::FilenamingConvention),
             "useNamingConvention" => Deserializable::deserialize(value, "options", diagnostics)