+++ crates/biome_service/src/configuration/parse/json/css/mod.rs
@@ -148,7 +148,7 @@ impl DeserializationVisitor for CssFormatterVisitor {
                 "indentSize" => {
                     result.indent_width =
                         Deserializable::deserialize(&value, &key_text, diagnostics);
-                    diagnostics.push(DeserializationDiagnostic::new_deprecated(
+                    diagnostics.push(DeserializationDiagnostic::new_deprecated_use_instead(
                         &key_text,
                         key.range(),
                         "css.formatter.indentWidth",